+++ example/components/my-advanced-greeting.js
@@ -3,25 +3,28 @@ import {
     useState,
 } from "/lib/function-web-components/index.js";
 import { html, render } from "/lib/lit-html/lit-html.js";
+import "./my-prompt-input.js";
 
 function myAdvancedGreeting({ givenName }) {
     const [surname, setSurname] = useState("...");
-    const askForSurname = () => {
-        const response = prompt("Enter surname:");
-        if (response) {
-            setSurname(response);
-        }
+    const onSurnameChange = e => {
+        setSurname(e.detail);
     };
 
     return html`
         <p>
-            Hello ${givenName} ${surname}.
-            <button @click=${askForSurname}>Change surname</button>
+            Hello ${givenName}
+            <my-prompt-input
+                value="${surname}"
+                .promptBuilder=${() => `Current name: ${givenName} ${surname}`}
+                @change=${onSurnameChange}
+            ></my-prompt-input
+            >.
         </p>
     `;
 }
 
 customElements.define(
     "my-advanced-greeting",
-    makeWebComponent(myAdvancedGreeting, { props: ["givenName"], render })
+    makeWebComponent(myAdvancedGreeting, { attrs: ["givenName"], render })
 );